[FR-141] fix: 마이페이지 회원정보 수정 비밀번호 검증 코드 위치 수정
- 기존에는 user 도메인에 있었는데 my 도메인으로 이동
